pyproject.toml
[project]
name = "project-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.22.1",
    "google-auth>=2.47.0",
    "google-genai>=1.59.0",
    "python-dotenv>=1.2.1",
]
